Add project description, URL, inception year, and SCM info to POM (#33)
I am hoping that, in addition to making the POM better, this will
make Sonar update its links and project description which are currently
incorrect...not sure where exactly it got them from, e.g. the URL that
Sonar says is the homepage for kiwi-beta is:

https://github.com/kiwiproject/kiwi-parent/kiwi-beta

which is not even valid.