Update ManagedCamelContext to allow giving a descriptive name (#107)
If an application has more than one CamelContext, then being able to
give them names is useful when viewing log messages that contain
the name. The existing constructor creates a default name that
includes the current timestamp. It seems highly unlikely that two
CamelContexts would be created at the exact same instance in the same
JVM, so I used System#currentTimeMillis as the timestamp.

Also, add argument checking to the 2-arg constructor and change logging
to use CamelContext#getRouteSize since the #getRoutes method creates
a copy of the routes inside a synchronized block, etc. and all we need
is the size for this replacement value.