Replace test causing stack overflow in JarManifestsTest (#127)
Replace the test that was causing StackOverflowException in
JarManifestsTest with simpler ones that do not need to perform
any shenanigans with the SecurityManager, which is deprecated
for removal in JDK 17. Refactored existing code to permit easy
"mocking" of the ProtectionDomain and CodeSource using old
school concrete inheritance. In future we might need to revisit
the exception thrown by getProtectionDomain() since I would
assume it won't throw a SecurityException once the SecurityManager
is removed from Java. That will probably be a while, however, so
not worrying too much about it right now.

Fixes #126