Add AccessedViaReflection annotation (#154)
This annotation can be used to indicate that a constructor, method, or
field is accessed by reflection, to make their usage more clear. For example
if a method is only used in unit tests, someone might think it can be safely removed. Adding this annotation would help in that situation.