Refactor RuntimeSQLExceptionTest to use chaining (#168)
By changing the assertions on the RuntimeSQLException to chain, we
can avoid the 'shenanigans' og having to declare it as a Throwable
to avoid ambiguity in Assertions#assertThat. Plus, it's cleaner code
that better communicates the intent. And, as much as I love to write
comments explaining arcane weirdness, it's better not to need them
at all.