Increase timeout in ProcessHelpersTest
Occasionally I see timeouts in shouldHandleProcessExceptionsGracefully
in ProcessHelpersTest, mainly while in Gitpod inside a browser. The current
timeout is 75 milliseconds. I am guessing that the Mockito spy is causing
the timeouts, since it sometimes takes time to initialize Mockito (I don't
know why it takes a long time, but this is my observed behavior). For
now, "fix" the problem by increasing the timeout to 250 milliseconds.