Add DefaultingFileAppenderFactory (#192)
DefaultingFileAppenderFactory is a Dropwizard AppenderFactory extending
FileAppenderFactory and provides default values for currentLogFilename,
archivedLogFilenamePattern, and logFormat if not explicitly provided.

A META-INF service file is also provided containing the FQCN. This must
be combined (e.g. using Maven Shade and ServicesResourceTransformer)
with Dropwizard's out-of-box file, as noted in the javadoc.