Add KiwiFutures (#217)
Contains two methods that replicate the functionality of several methods
that were added to Future in JDK 19, namely resultNow and exceptionNow.
Here, they are obviously not added to Future, since it's a JDK interface that
we cannot change. Instead they are implemented as static utility methods
that accept a Future argument.