Remove commented-out 'nowarn' option from kotlin-maven-plugin (#226)
When true, the 'nowarn' option suppresses the (kotlin) compiler from
displaying warnings during compilation. The default value is false,
meaning it will display compiler warnings.

I've had this commented out for a while. Since warnings have not been a
problem, there's no point in having this option sitting here in the
POM, commented out,m and with an annoying ever-present TODO. So,
removing it.