Add emptyArray and newArray to KiwiReflection2 (#276)
These methods allow you to easily create an array of a given type
without needing to cast. One creates an empty array while the other
creates an array with a given length, and which contains only null
values.

The javadocs and initial version of the tests were "written" by ChatGPT.
I modified them slightly, and added a few more tests. For example, a
ValueSource can be used instead of a MethodSource when there is only one
test parameter. I also had to slightly correct the text in the throws
explanations in the javadoc, plus I added additional text including the
"see" references and the implNote text.