Allow Qodana to write security events (#403)
Hopefully this will fix the "Resource not accessible by integration" error.

The full error says:

This run of the CodeQL Action does not have permission to
access Code Scanning API endpoints. As a result, it will not be
opted into any experimental features. This could be because the
Action is running on a pull request from a fork. If not, please
ensure the Action has the 'security-events: write' permission.
Details: Resource not accessible by integration