Add new methods to Timing that don't require an explicit StopWatch (#444)
Add several new methods to Timing which do not require a StopWatch. This
required new "result" classes, because any exceptions must be caught
instead of thrown, so that callers can retrieve elapsed time even if an
exception occurs.

The new methods and result types provide nanosecond timing instead of
just milliseonds, like the original methods in this class do.

The new result types are the sealed interface TimedWithError and its two
implementations, TimedWithErrorOrResult and TimedWithErrorNoResult.

The new timeWithResult and timeNoResult methods are annotated with the
error prone CheckReturnValue annotation, to make it clear that it must
be inspected to get the elapsed time and result or exception (though it
seems like it should be pretty obvious).

Changed the original Timed interfaced to be sealed and to permit only
TimedWithResult and TimedNoResult, since no one else should be creating
subclasses.

Also made a few minor javadoc changes to the original methods.