Add JarManifestHelper as an instance-based alternative to JarManifests (#465)
* Add JarManifestHelper, which is an instance-based class that contains
much the same functionality as JarManifests. But instead of having only
static methods, it only contains instance methods which make this much
easier to mock in unit tests of classes that need to perform JAR
manifest lookups.
* The methods in JarManifestHelper also return records containing more
information about the lookup result, instead of returning Optional or
throwing an exception. This allows callers to inspect the result and
decide what action to take based on more detailed information.
* Add Javadocs to JarManifests.