KiwiSlf4J enhancements to accept level names and convert String to Level (#503)
* Add toLevelIgnoreCase method to get the Level object that matches a
string whose characters have all been converted to uppercase.
* Add overloads for the isEnabled and log methods that accept a String
for the level name instead of the Level object. The level name is
case-insensitive, so that "warn", "Warn", "WARN", and even "WaRn" all
match Level.WARN.
* For the tests, I chose to copy and modify the existing implementations
instead of changing the tests to always call the methods that accept a
String (since they convert the level name to Level and then delegate to
the ones that accept Level). I made this choice to keep the tests
completely independent, at the cost of some duplicate test code.