Reduce WARN/ERROR log level to DEBUg in JarManifests
The methods where these logs occur return a ManifestLookupResult.
Callers can (and should) inspect the value and decide how to handle
the result. They can choose to log at a higher level like WARN or
ERROR of they want, but we should not log a WARN/ERROR and also
return a result that contains the error.

Closes #505