Add ability to escape a response entity when the content type HTML (#535)
Add KiwiHttpResponses2 with a single method: htmlEscape.

Add KiwiResponses2 with a single method: htmlEscapeEnity.

KiwiResponses2 handles HTML escaping when using Jakarta REST. You pass a
Jakarta REST Response to htmlEscape.

It delegates to KiwiHttpResponses2, which handles HTML escaping in a
generic manner. The htmlEscape method accepts two Strings: the
(response) entity to be HTML-escaped, and the media type.

In both methods, if the media type is not HTML, then no escaping is
performed.

Added jersey-client and the OkHttp mockwebserver as test-scoped
dependencies.

Closes #523